producer/requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
kafka-python==2.2.2
requests==2.32.4
urllib3==2.5.0
